Added the data cleaning Python file
This is a simple Python file that I created and used to clean up the Wikipedia URL in the original constructors table. Some links are not the final link (i.e., they get redirected to other addresses when pasted in browser), so this little program gets the final url for these pages.